requirements.txt
flask==3.1.2
flask_sqlalchemy==3.0.3
flask_jwt_extended==4.4.4
qrcode==7.3
pandas==2.1.1
python-docx==0.8.11
reportlab==4.1.5
